pyproject.toml
[tool.poetry]
name = "dbt-af"
version = "0.3.4"
description = "Distibuted dbt runs on Apache Airflow"
authors = ["Nikita Yurasov <nikitayurasov@toloka.ai>"]
readme = "README.md"
packages = [
    { include = "dbt_af" },
    { include = "dbt_af_functional_tests" },
    { include = "scripts" },
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
attrs = "^23.1"
apache-airflow = ">=2.6,<2.8.0"
pydantic = "^1.10.7"
apache-airflow-providers-cncf-kubernetes = ">=7.0.0"
cachetools = "^5.3"
typer = "^0.9"
dbt-core = "^1.7"
pendulum = "^2"

pytest = { version = "*", optional = true }
pytest-env = { version = "*", optional = true }
pytest-socket = { version = "*", optional = true }
pytest-mock = { version = "*", optional = true }

airflow-mcd = { version = "^0.2", optional = true }
pycarlo = { version = "^0.8", optional = true }

dbt-postgres = { version = "^1.7", optional = true }

[tool.poetry.group.dev.dependencies]
isort = "^5.12"
black = { extras = ["d"], version = "^23.7" }
twine = "*"
pytest = "^8"
pytest-env = "*"
pytest-socket = "*"
pytest-mock = "*"
dbt-postgres = "^1.7"
dbt-databricks = "^1.7"
databricks-cli = "^0.18"

[tool.poetry.extras]
mcd = ["airflow-mcd", "pycarlo"]
tests = ["pytest", "pytest-env", "pytest-socket", "pytest-mock"]
minidbt = []
examples = ["dbt-postgres"]

[tool.poetry.scripts]
dbt-af-manifest-tests = { callable = "dbt_af_functional_tests:cli", extras = ["tests"] }
mini_dbt_project_generator = { callable = "scripts.mini_dbt_project_generator:cli", extras = ["minidbt"] }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.isort]
line_length = 120
profile = "black"
